provider/pyproject.toml
[project]
name = "provider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "pact-python>=2.3.1",
    "prometheus-client>=0.21.1",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.optional-dependencies]
dev = [
    "gitpython>=3.1.44",
    "pytest>=7.0.0",
    "tomli>=2.2.1",
]
